fix noise implementation

+++ datapack-loading/src/main/java/net/minestom/vanilla/datapack/worldgen/noise/NormalNoise.java
@@ -5,7 +5,7 @@ import net.minestom.vanilla.datapack.worldgen.random.WorldgenRandom;
 
 public class NormalNoise implements Noise {
 
-    public record Config(double firstOctave, DoubleList amplitudes) {
+    public record Config(int firstOctave, DoubleList amplitudes) {
     }
 
     private static final double INPUT_FACTOR = 1.0181268882175227;
@@ -16,7 +16,7 @@ public class NormalNoise implements Noise {
     public final double maxValue;
 
     public NormalNoise(WorldgenRandom random, Config config) {
-        double firstOctave = config.firstOctave();
+        int firstOctave = config.firstOctave();
         DoubleList amplitudes = config.amplitudes();
         this.first = new PerlinNoise(random, firstOctave, amplitudes);
         this.second = new PerlinNoise(random, firstOctave, amplitudes);
